+++ gradle/gradle-plugins/src/main/kotlin/org/http4k/internal/ModuleLicense.kt
@@ -1,6 +1,6 @@
 package org.http4k.internal
 
-import gradle.kotlin.dsl.accessors._45fd76dc729b81c2a4255a063893a6f3.sourceSets
+import gradle.kotlin.dsl.accessors._07863873f264822bc766251a55df02df.sourceSets
 import org.gradle.api.Project
 import org.gradle.api.tasks.bundling.Jar
 import org.gradle.kotlin.dsl.named
@@ -8,12 +8,12 @@ import org.gradle.kotlin.dsl.named
 enum class ModuleLicense(val commonName: String, val url: String, val licenseDir: String) {
     Apache2(
         "Apache-2.0",
-        "http://www.apache.org/licenses/LICENSE-2.0",
+        "https://www.apache.org/licenses/LICENSE-2.0",
         "."
     ),
     Http4kCommercial(
         "http4k Commercial License",
-        "http://http4k.org/commercial-license",
+        "https://http4k.org/commercial-license",
         "./pro"
     )
 }